API Functions
Wrote the functions of the API that will be used in the implementation. We need to be able to make a specific pond in order to implement the Implementation but it should just be a matter of taking the pond and running the functions for each time step and having it output the data. Also made a few small changes to the Fish and Pond files. Added an export to pond and made the prey attribute of a fish be a list of populations rather than a list of fish types.